Mais um exemplo

+++ Listas/[4]gulosos-stl/third-party-software/third-party-software.cpp
@@ -6,17 +6,17 @@ using namespace std;
 
 int main()
 {
-	int n; // 1 ≤ n≤ 200.000
-	int m; // 1 ≤ m≤ 10^9 
+	long n; // 1 ≤ n ≤ 200.000
+	long m; // 1 ≤ m ≤ 10^9 
 
 	cin >> n >> m;
 
-	for (int i = 0; i < n; i++)
+	for (long i = 0; i < n; i++)
 	{
-		int min, max;
+		long min, max;
 		cin >> min >> max;
 
-		
+
 	}
